Move SingleEntityPhiTensor
